requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.26.0
pydantic==2.6.1
numpy==1.24.3
transformers==4.37.2
torch==2.2.0
python-multipart==0.0.9
google-generativeai==0.3.2
python-dotenv==1.0.0
sentencepiece==0.1.99
protobuf==4.25.1 